Use different namespace for collections and documents' data
A document may contain a key and a sub-collection with the same name:
they should not conflict.

By having separate namespaces for data and sub-collections, this also
fixes #90: indeed, updating a document's data without setting the merge
flag won't remove sub-collections.

One limitation is the `dump()` function of the `MockFirestoreInstance`
class: because it rerturns a single tree, precedence is given to the
sub-collection is there is a name conflict between a sub-collection and
a key in the data.

V2: Address comments from @atn832:
      * Added comments to the dump() function's new loop.
      * Used more variables to make the code more readable.
      * Declared varialbes as final when relevant.
      * Added a testcase; see previous commit.

Fixes #90.