Fix bug with replacing non-map values with Map
Previously the code worked under the assumption that if the new value for a given key was a map, and the key existed in the document, then the existing value was also a map. This is not always the case. This change updates the code to check whether the value is a map, and update it to one, before continuing to apply values.